# FarmlyCore - Copy/FarmlyCore.Api/wwwroot/docs/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "FarmlyCore",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:5001"
    }
  ],
  "paths": {
...
    "/customer-access/customer/find": {
      "post": {
        "tags": [
          "customer"
        ],
        "operationId": "Customer_FindCustomers",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindCustomersRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/customer-access/customer": {
      "post": {
        "tags": [
          "customer"
        ],
        "operationId": "Customer_CreateCustomer",
        "requestBody": {
          "x-name": "customer",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
    "/customer-access/customer/find-users": {
      "post": {
        "tags": [
          "customer"
        ],
        "operationId": "Customer_FindOrder",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindUsersRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomerAddressDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "integer",
            "format": "int32"
          },
          "fkCustomerId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "CustomerDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "companyName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "customerType": {
            "$ref": "#/components/schemas/CustomerTypeDto"
          },
          "bankGiro": {
            "type": "string"
          },
          "orgNumber": {
            "type": "string"
          },
          "customerAddresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerAddressDto"
            }
          }
        }
      },
      "CustomerTypeDto": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Farmer",
          "Restaurant"
        ],
        "enum": [
          "Farmer",
          "Restaurant"
        ]
      },
      "FindCustomersRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "customerType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerTypeDto"
              }
            ]
          },
          "companyName": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string"
          },
          "customerId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "FindUsersRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "credentials": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "customerId": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}